feat: rename files from slackAlert to alert
