+++ dashboard/components/inventory/components/view/alerts/InventoryViewAlertsEditAlert.tsx
@@ -4,8 +4,8 @@ import Grid from '../../../../grid/Grid';
 import ArrowLeftIcon from '../../../../icons/ArrowLeftIcon';
 import Input from '../../../../input/Input';
 import { ToastProps } from '../../../../toast/hooks/useToast';
-import useEditAlerts from './hooks/useEditSlackAlerts';
-import { AlertMethod, Alert } from './hooks/useSlackAlerts';
+import useEditAlerts from './hooks/useEditAlerts';
+import { AlertMethod, Alert } from './hooks/useAlerts';
 import settingsService from '../../../../../services/settingsService';
 import { useState } from 'react';
 import LoadingSpinner from '../../../../icons/LoadingSpinner';
