+++ dashboard/components/inventory/components/view/alerts/hooks/useAlerts.tsx
