+++ dashboard/components/inventory/components/view/alerts/hooks/useEditAlerts.tsx
@@ -1,7 +1,7 @@
 import { FormEvent, useState } from 'react';
 import settingsService from '../../../../../../services/settingsService';
 import { ToastProps } from '../../../../../toast/hooks/useToast';
-import { AlertMethod, Alert } from './useSlackAlerts';
+import { AlertMethod, Alert } from './useAlerts';
 
 type AlertType = 'BUDGET' | 'USAGE';
